requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.14.2
bleach==6.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
colorlog==6.9.0
crashtest==0.4.1
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
docopt==0.6.2
dulwich==0.22.7
executing==2.2.1
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
frozenlist==1.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keyring==25.6.0
magic-filter==1.0.12
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mistune==3.1.4
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.6.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.3
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pbs-installer==2025.2.12
pexpect==4.9.0
pickleshare==0.7.5
pip-check-reqs==2.5.5
pipreqs==0.5.0
pkginfo==1.12.1.1
platformdirs==4.3.6
poetry==2.1.1
poetry-core==2.1.1
prompt_toolkit==3.0.52
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
RapidFuzz==3.12.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
tinycss2==1.4.0
tomlkit==0.13.2
tornado==6.5.2
traitlets==5.14.3
trove-classifiers==2025.2.18.16
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
virtualenv==20.29.2
wcwidth==0.2.14
webencodings==0.5.1
xattr==1.1.4
xlsxwriter==3.2.9
yarg==0.1.9
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
psutil~=7.1.0

openpyxl~=3.1.5
transliterate~=1.10.2